breakout IQueueMessageAttachmentProvider + max queue size for SB
