+++ knightbus-azureservicebus/src/KnightBus.Azure.ServiceBus.Management/KnightBus.Azure.ServiceBus.Management.csproj
@@ -10,7 +10,7 @@
     <PackageIconUrl>https://raw.githubusercontent.com/BookBeat/knightbus/master/documentation/media/images/knighbus-64.png</PackageIconUrl>
     <PackageIcon>knighbus-64.png</PackageIcon>
     <RepositoryUrl>https://github.com/BookBeat/knightbus</RepositoryUrl>
-    <Version>4.0.0$(VersionSuffix)</Version>
+    <Version>4.1.0$(VersionSuffix)</Version>
     <PackageTags>knightbus;azure servicebus;amqp;queues;messaging</PackageTags>
   </PropertyGroup>
