+++ knightbus-azurestorage/src/KnightBus.Azure.Storage.Management/KnightBus.Azure.Storage.Management.csproj
@@ -9,7 +9,7 @@
     <PackageIconUrl>https://raw.githubusercontent.com/BookBeat/knightbus/master/documentation/media/images/knighbus-64.png</PackageIconUrl>
     <PackageIcon>knighbus-64.png</PackageIcon>
     <RepositoryUrl>https://github.com/BookBeat/knightbus</RepositoryUrl>
-    <Version>4.0.0$(VersionSuffix)</Version>
+    <Version>4.1.0$(VersionSuffix)</Version>
     <PackageTags>knightbus;azure storage;blob;queues;messaging</PackageTags>
   </PropertyGroup>
